Renamed directory, moved some classes to own directory
