+++ src/base/DataError.ts
@@ -1,11 +1,11 @@
 /**
  * An error that may be thrown to indicate that input
- * data was invalid. 
- * 
+ * data was invalid.
+ *
  * This may refer to buffers that have been expected to
  * contain a certain type of data, but did not. For
  * example, a buffer that looked like a GZIpped buffer,
- * but turned out to be invalid, or a buffer that 
+ * but turned out to be invalid, or a buffer that
  * looked like it contained valid JSON, but did not.
  *
  * @internal
