+++ src/contentProcessing/GltfPipelineLegacy.ts
