+++ src/contentProcessing/GtlfUtilities.ts
