+++ src/implicitTiling/ImplicitTilingError.ts
@@ -4,7 +4,7 @@
  *
  * This may be thrown by methods that create the convenience classes
  * for this package, when the given resources are not valid.
- * 
+ *
  * @internal
  */
 export class ImplicitTilingError extends Error {
