+++ src/implicitTiling/SubtreeInfo.ts
@@ -6,7 +6,7 @@ import { AvailabilityInfo } from "./AvailabilityInfo";
  *
  * It offers the availability information for tiles, child
  * subtrees, and contents, as `AvailabilityInfo` objects.
- * 
+ *
  * @internal
  */
 export interface SubtreeInfo {
