+++ src/implicitTiling/SubtreeInfos.ts
@@ -10,7 +10,7 @@ import { TileImplicitTiling } from "../structure/TileImplicitTiling";
 
 /**
  * Methods to create `SubtreeInfo` instances.
- * 
+ *
  * @internal
  */
 export class SubtreeInfos {
