+++ src/metadata/ClassProperties.ts
@@ -7,7 +7,7 @@ import { ClassProperty } from "../structure/Metadata/ClassProperty";
 
 /**
  * Utility methods related to `ClassProperty` objects
- * 
+ *
  * @internal
  */
 export class ClassProperties {
