+++ src/spatial/OctreeCoordinates.ts
@@ -4,7 +4,7 @@ import { TreeCoordinates } from "./TreeCoordinates";
 
 /**
  * An implementation of `TreeCoordinates` for octrees
- * 
+ *
  * @internal
  */
 export class OctreeCoordinates implements TreeCoordinates {
