+++ src/spatial/QuadtreeCoordinates.ts
@@ -4,7 +4,7 @@ import { TreeCoordinates } from "./TreeCoordinates";
 
 /**
  * An implementation of `TreeCoordinates` for octrees
- * 
+ *
  * @internal
  */
 export class QuadtreeCoordinates implements TreeCoordinates {
