+++ src/spatial/Quadtrees.ts
@@ -1,20 +1,20 @@
-import { OctreeCoordinates } from "./OctreeCoordinates";
+import { QuadtreeCoordinates } from "./QuadtreeCoordinates";
 
 /**
- * Methods related to octrees
- * 
+ * Methods related to quadtrees.
+ *
  * @internal
  */
-export class Octrees {
+export class Quadtrees {
   /**
-   * Computes the number of nodes of an octree with the given number of
+   * Computes the number of nodes of a quadtree with the given number of
    * levels.
    *
    * @param levels - The number of levels
    * @returns The number of nodes
    */
   static computeNumberOfNodesForLevels(levels: number) {
-    return ((1 << (levels * 3)) - 1) / (8 - 1);
+    return ((1 << (levels * 2)) - 1) / (4 - 1);
   }
 
   /**
@@ -25,28 +25,25 @@ export class Octrees {
    */
   static *coordinatesForLevel(level: number) {
     const size = 1 << level;
-    for (let z = 0; z < size; z++) {
-      for (let y = 0; y < size; y++) {
-        for (let x = 0; x < size; x++) {
-          yield new OctreeCoordinates(level, x, y, z);
-        }
+    for (let y = 0; y < size; y++) {
+      for (let x = 0; x < size; x++) {
+        yield new QuadtreeCoordinates(level, x, y);
       }
     }
   }
 
   /**
    * Returns whether the given coordinates are valid. This means
-   * that the x, y, and z components are in the range that is
+   * that the x and y components are in the range that is
    * determined by the level of the coordinates.
    *
    * @param c - The coordinates
    * @returns Whether the coordinates are valid
    */
-  static isValid(c: OctreeCoordinates) {
+  static isValid(c: QuadtreeCoordinates) {
     const level = c.level;
     const x = c.x;
     const y = c.y;
-    const z = c.z;
     if (level < 0) {
       return false;
     }
@@ -55,9 +52,6 @@ export class Octrees {
       return false;
     }
     if (y < 0 || y >= size) {
-      return false;
-    }
-    if (z < 0 || z >= size) {
       return false;
     }
     return true;
