+++ src/spatial/TreeCoordinates.ts
@@ -1,6 +1,6 @@
 /**
  * An interface for coordinates within a tree structure.
- * 
+ *
  * @internal
  */
 export interface TreeCoordinates {
