+++ src/traversal/ImplicitTraversedTile.ts
@@ -11,7 +11,8 @@ import { Content } from "../structure/Content";
 import { MetadataEntity } from "../structure/MetadataEntity";
 import { TileImplicitTiling } from "../structure/TileImplicitTiling";
 
-import { TreeCoordinates } from "../implicitTiling/TreeCoordinates";
+import { TreeCoordinates } from "../spatial/TreeCoordinates";
+
 import { ImplicitTilingError } from "../implicitTiling/ImplicitTilingError";
 import { ImplicitTilings } from "../implicitTiling/ImplicitTilings";
