+++ src/traversal/SubtreeModel.ts
@@ -3,7 +3,7 @@ import { SubtreeMetadataModel } from "./SubtreeMetadataModel";
 
 /**
  * An interface that summarizes the information for a subtree.
- * 
+ *
  * @internal
  */
 export interface SubtreeModel {
